Package refactoring and wp generation update.
