+++ MyShop - Warehouse Module/com/myshop/warehouse/generators/Generator.java
@@ -1,0 +1,6 @@
+package com.myshop.warehouse.generators;
+
+public interface Generator {
+
+	Object generate();
+}
