+++ MyShop - Warehouse Module/com/myshop/warehouse/util/GetLastIndexCreated.java
@@ -1,0 +1,9 @@
+package com.myshop.warehouse.util;
+
+public class GetLastIndexCreated {
+
+	public int get() {
+		return -1;
+	}
+
+}
